# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DDpay https://github.com/zkeai/DDPay",
        "title": "DDPay API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:2900",
    "basePath": "/api/v1",
    "paths": {
...
        "/order/create-transaction": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "创建订单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        },
        "/pay/status": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "获取当前订单状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单key",
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/conf.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "conf.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.OrderReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "notify_url": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "redirect_url": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "trade_type": {
                    "type": "string"
                }
            }
        },
...
    }
}